Abstracted NotationParser class out into separate file
