+++ tests/test_music/test_notationparser.py
@@ -1,0 +1,10 @@
+import pytest
+
+from chordparser.music.notationparser import NotationParser
+
+
+class TestNNPGetRegexGroupsCount:
+    def test_correct_number_of_groups(self):
+        parser = NotationParser()
+        parser._pattern = "(a)(b)"
+        assert 2 == parser.get_regex_groups_count()
